refactor: add error service in context (#531)
