+++ src/routes/stats.js
@@ -1,7 +1,7 @@
 const _ = require('lodash');
 const auth = require('../services/auth');
 const logger = require('../services/logger');
-const error = require('../services/error');
+const error = require('../utils/error');
 const path = require('../services/path');
 const StatSerializer = require('../serializers/stat');
 const Schemas = require('../generators/schemas');
